requirements.txt
# Core LangGraph dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langchain-tavily>=0.1.0

# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
